阿里云分支

+++ hs-item/hs-item-interface/src/main/java/com/hsource/item/entity/Invitation.java
@@ -51,7 +51,7 @@ public class Invitation extends BaseEntity {
      * 喜欢数
      */
     @TableField("like_num")
-    private long likeNum;
+    private Integer likeNum;
 
     /**
      * 内容
